Changed: Minors server changes from Hg

+++ src/gpm_service/messages.cpp
@@ -14,6 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+#include "stdpch.h"
 
 // include files
 #include "nel/misc/types_nl.h"
@@ -816,7 +817,7 @@ void cbVisionRequest( NLNET::CMessage& msgin, const std::string &serviceName, NL
 	CMessage	msgout("VISION_ANSWER");
 	msgout.serial(rid);
 	uint	i;
-	const uint size = entities.size();
+	const uint size = (uint)entities.size();
 	for ( i = 0 ; i < size ; ++i)
 		msgout.serial(entities[i].first, entities[i].second);
